python/requirements.txt
# mjpegZero — Python dependencies for reference encoder and verification
#
# Install:  pip install -r python/requirements.txt

numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0
